Code Refactor
Seperate UI events into separate module, reorganize other events logically, add static routines
