+++ WebLoader.csproj
@@ -8,7 +8,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <StartupObject>WebLoader.Program</StartupObject>
     <Title>Simple Browser</Title>
-    <Version>1.2.5</Version>
+    <Version>1.2.7</Version>
     <Product>Simple Browser</Product>
     <Copyright>(c) 2024 J.D. Chapman (All Rights Reserved)</Copyright>
     <PackageIcon>SimpleBrowser.ico</PackageIcon>
